+++ CHANGELOG.md
@@ -25,6 +25,9 @@ CHANGELOG
 - `pulumi stack` now renders the stack as a tree view.
   [#3430](https://github.com/pulumi/pulumi/pull/3430)
 
+- Support for lists and maps in config.
+  [#3342](https://github.com/pulumi/pulumi/pull/3342)
+
 ## 1.4.0 (2019-10-24)
 
 - `FileAsset` in the Python SDK now accepts anything implementing `os.PathLike` in addition to `str`.